+++ packages/flame_jenny/jenny/lib/src/structure/expressions/functions/ceil.dart
@@ -1,5 +1,5 @@
 import 'package:jenny/src/structure/expressions/expression.dart';
-import 'package:jenny/src/structure/expressions/functions/_utils.dart';
+import 'package:jenny/src/structure/expressions/functions/_common.dart';
 import 'package:jenny/src/yarn_project.dart';
 
 /// Function `ceil(x)` will round `x` up towards positive infinity.
