+++ packages/flame_jenny/jenny/lib/src/structure/expressions/functions/random_range.dart
@@ -1,6 +1,6 @@
 import 'package:jenny/src/errors.dart';
 import 'package:jenny/src/structure/expressions/expression.dart';
-import 'package:jenny/src/structure/expressions/functions/_utils.dart';
+import 'package:jenny/src/structure/expressions/functions/_common.dart';
 import 'package:jenny/src/yarn_project.dart';
 
 /// Function `random_range(a, b)` returns a random integer between `a` and `b`,
