+++ packages/flame_jenny/jenny/lib/src/structure/expressions/functions/round_places.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:jenny/src/structure/expressions/expression.dart';
-import 'package:jenny/src/structure/expressions/functions/_utils.dart';
+import 'package:jenny/src/structure/expressions/functions/_common.dart';
 import 'package:jenny/src/yarn_project.dart';
 
 /// Function `round_places(x, n)` will round `x` to `n` decimal places.
